+++ src/tasks.rs
@@ -464,9 +464,9 @@ impl Task {
     /// returns: HashMap<String, String, RandomState>
     fn get_vars(
         &self,
-        env: &HashMap<String, serde_yaml::Value>,
-    ) -> HashMap<String, serde_yaml::Value> {
-        let mut new_vars: HashMap<String, serde_yaml::Value> = self.common.vars.clone();
+        env: &HashMap<String, serde_json::Value>,
+    ) -> HashMap<String, serde_json::Value> {
+        let mut new_vars: HashMap<String, serde_json::Value> = self.common.vars.clone();
         for (key, val) in env {
             new_vars.entry(key.clone()).or_insert_with(|| val.clone());
         }
@@ -533,7 +533,7 @@ impl Task {
         args: &ArgsContext,
         mom_file: &MomFile,
         env: &HashMap<String, String>,
-        vars: &HashMap<String, serde_yaml::Value>,
+        vars: &HashMap<String, serde_json::Value>,
     ) -> tera::Context {
         let mut context = tera::Context::new();